move CPU stuff to `cpu` module
