+++ src/cpu/execute.rs
@@ -1,6 +1,6 @@
-use crate::fetch::Fetch;
-use crate::registers::{flags, Alu};
-use crate::system::{InterruptHandler, MemoryIO, Stack, System};
+use crate::cpu::fetch::Fetch;
+use crate::cpu::registers::{flags, Alu};
+use crate::cpu::{InterruptHandler, MemoryIO, Stack, System};
 
 pub trait Execute {
   /// Execute the given opcode, returning either the number of cycles used or an error.
