+++ src/systems/brooke.rs
@@ -1,8 +1,8 @@
+use crate::cpu::System;
 use crate::memory::{ActiveInterrupt, Memory, SystemInfo};
 use crate::memory::{BlockMemory, BranchMemory};
 use crate::platform::PlatformProvider;
 use crate::roms::RomFile;
-use crate::system::System;
 use crate::systems::SystemFactory;
 use std::io::Write;
 use std::sync::Arc;
