+++ src/systems/easy.rs
@@ -1,8 +1,8 @@
+use crate::cpu::System;
 use crate::keyboard::KeyPosition;
 use crate::memory::{ActiveInterrupt, BlockMemory, BranchMemory, Memory, SystemInfo};
 use crate::platform::{Color, PlatformProvider, WindowConfig};
 use crate::roms::RomFile;
-use crate::system::System;
 use crate::systems::SystemFactory;
 use std::sync::Arc;
